qa: apply coding standard and ignore some psalm issues regarding `class_alias`
