+++ phpcs.xml.dist
@@ -16,6 +16,8 @@
     <file>bin</file>
     <file>src</file>
     <file>test</file>
+    <!-- psr/container v2 exclusion -->
+    <exclude-pattern>src/AbstractUntypedContainerImplementation.php</exclude-pattern>
     <exclude-pattern>test/TestAsset/laminas-code/*.php</exclude-pattern>
 
     <!-- Include all rules from the Laminas Coding Standard -->
