+++ psalm.xml.dist
@@ -50,16 +50,16 @@
         <MethodSignatureMismatch>
             <errorLevel type="suppress">
                 <!-- <editor-fold desc="psr/container v2 hacks"> -->
-                <file name="src/TypedAbstractContainerImplementation.php"/>
-                <file name="src/UntypedAbstractContainerImplementation.php"/>
+                <file name="src/AbstractTypedContainerImplementation.php"/>
+                <file name="src/AbstractUntypedContainerImplementation.php"/>
                 <!-- </editor-fold> -->
             </errorLevel>
         </MethodSignatureMismatch>
         <TypeDoesNotContainType>
             <errorLevel type="suppress">
                 <!-- <editor-fold desc="psr/container v2 hacks"> -->
-                <file name="src/TypedAbstractContainerImplementation.php"/>
-                <file name="src/UntypedAbstractContainerImplementation.php"/>
+                <file name="src/AbstractTypedContainerImplementation.php"/>
+                <file name="src/AbstractUntypedContainerImplementation.php"/>
                 <!-- </editor-fold> -->
             </errorLevel>
         </TypeDoesNotContainType>
